Decoupled Network specific logic from RadixClient.java
* Will make JUnit testing a lot easier
* Will make networking debugging a lot easier
* Will allow RadixClient to use other protocols other than websockets
